@@ -1,9 +1,9 @@
 {
-  "name": "electric-sql-wa-sqlite-example",
-  "version": "0.7.0",
+  "name": "supabase-basic",
   "author": "ElectricSQL",
   "license": "Apache-2.0",
-  "main": "builder.js",
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "backend:start": "cd ./backend && docker compose up",
     "backend:stop": "cd ./backend && docker compose stop",
@@ -12,12 +12,10 @@
     "client:generate": "npx electric-sql generate --service http://localhost:5133 --proxy postgresql://prisma:proxy_password@localhost:65432/postgres",
     "client:watch": "npm run client:generate --watch",
     "db:migrate": "npx pg-migrations apply --database postgresql://postgres:proxy_password@localhost:65432/postgres --directory ./db/migrations",
-    "build": "node copy-wasm-files.js && node builder.js",
-    "build-prod": "NODE_ENV=production npm run build",
-    "start": "SERVE=true npm run build"
-  },
-  "engines": {
-    "node": ">=16.11.0"
+    "dev": "node copy-wasm-files.js && vite",
+    "build": "node copy-wasm-files.js && tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
     "electric-sql": "^0.7.0",
@@ -29,16 +27,16 @@
     "@databases/pg-migrations": "^5.0.2",
     "@electric-sql/prisma-generator": "^1.1.0",
     "@prisma/client": "4.8.1",
-    "@types/node": ">=16.11.0",
-    "@types/react": "^18.0.18",
-    "@types/react-dom": "^18.0.11",
-    "esbuild": "^0.16.17",
-    "esbuild-plugin-inline-image": "^0.0.9",
-    "fs-extra": "^10.0.0",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "@vitejs/plugin-react": "^4.2.0",
+    "eslint": "^8.53.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.4",
     "prisma": "4.8.1",
-    "prompt": "^1.3.0",
-    "shelljs": "^0.8.5",
-    "tcp-port-used": "^1.0.2",
-    "typescript": "^4.4.3"
+    "typescript": "^5.2.2",
+    "vite": "^5.0.0"
   }
 }
