@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 
 import { LIB_VERSION } from 'electric-sql/version'
 import { makeElectricContext, useLiveQuery } from 'electric-sql/react'
@@ -6,7 +6,6 @@ import { genUUID, uniqueTabId } from 'electric-sql/util'
 import { ElectricDatabase, electrify } from 'electric-sql/wa-sqlite'
 
 import { authToken } from './auth'
-import { DEBUG_MODE, ELECTRIC_URL } from './config'
 import { Electric, Items as Item, schema } from './generated/client'
 
 import './Example.css'
@@ -24,8 +23,8 @@ export const Example = () => {
         auth: {
           token: authToken()
         },
-        debug: DEBUG_MODE,
-        url: ELECTRIC_URL
+        debug: import.meta.env.DEV,
+        url: import.meta.env.ELECTRIC_URL ?? 'http://localhost:5133'
       }
 
       const { tabId } = uniqueTabId()
